backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.6
SQLAlchemy==2.0.48
Flask-Mail==0.9.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
supabase==2.0.0
